+++ school-service/src/main/java/com/backfcdev/schoolservice/repository/ISchoolRepository.java
@@ -3,5 +3,5 @@ package com.backfcdev.schoolservice.repository;
 import com.backfcdev.schoolservice.model.School;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ISchoolRepository extends JpaRepository<School, Long> {
+public interface ISchoolRepository extends JpaRepository<School, Integer> {
 }
