+++ school-service/src/main/java/com/backfcdev/schoolservice/service/ISchoolService.java
@@ -9,8 +9,8 @@ import java.util.List;
 public interface ISchoolService {
     List<School> findAll();
     School save(School school);
-    School findById(Long id);
-    School update(Long id, School school);
-    void delete(Long id);
+    School findById(Integer id);
+    School update(Integer id, School school);
+    void delete(Integer id);
     FullSchoolResponse findSchoolWithStudents(Integer schoolId);
 }
